+++ src/features/pn/types/notificationRecipient.ts
@@ -1,0 +1,7 @@
+import { NotificationPaymentInfo } from "./notificationPaymentInfo";
+
+export type NotificationRecipient = {
+  taxId: string;
+  denomination: string;
+  payment: NotificationPaymentInfo;
+};
