+++ src/routers/features/siciliaVola/secured/index.ts
@@ -130,7 +130,7 @@ addHandler(
     }
     res.json({});
   },
-  2000
+  () => 2000
 );
 
 /**
